requirements.txt
flask
cryptography
python-dotenv
Werkzeug